Crate epaint_default_fonts

A library containing built-in fonts for epaint, embedded as bytes.

This is intended to be consumed through the epaint crate.

Constants

EMOJI_ICON

An experimental typeface that uses standardized UNICODE planes for icon fonts.

HACK_REGULAR

A typeface designed for source code.

NOTO_EMOJI_REGULAR

A typeface containing emoji characters as designed for the Noto font family.

UBUNTU_LIGHT

A typeface designed for use by Ubuntu.